ipv4 mode + overhaul
